+++ src/main/java/dev/willyelton/crystal_tools/gui/component/CrystalToolsButton.java
@@ -14,7 +14,6 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 import org.jetbrains.annotations.NotNull;
 
 public abstract class CrystalToolsButton extends Button {
-    public static final ResourceLocation SKILL_BUTTON_LOCATION = new ResourceLocation("crystal_tools", "textures/gui/furnace_button.png");
 
     protected final OnTooltip onTooltip;
 
@@ -27,13 +26,13 @@ public abstract class CrystalToolsButton extends Button {
     public void renderWidget(@NotNull GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick) {
         Minecraft minecraft = Minecraft.getInstance();
         Font font = minecraft.font;
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.setShaderTexture(0, SKILL_BUTTON_LOCATION);
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, this.alpha);
+//        RenderSystem.setShader(GameRenderer::getPositionTexShader);
+//        RenderSystem.setShaderTexture(0, SKILL_BUTTON_LOCATION);
+//        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, this.alpha);
         // need to override this based on my button texture
         int textureY = this.getTextureY(this.isHoveredOrFocused());
         RenderSystem.enableBlend();
-        RenderSystem.defaultBlendFunc();
+//        RenderSystem.defaultBlendFunc();
         RenderSystem.enableDepthTest();
         this.blitButton(guiGraphics, textureY);
         // pretty sure does nothing TODO: CHECK